+++ src/RcsbSeacrh/FacetStore/DepositionGroupFacetStore.ts
@@ -50,7 +50,7 @@ class DepositionGroupFacetStore implements FacetStoreInterface{
         cloneDeep<FacetMemberInterface>(CHEM_COMP_FACET)
     ];
 
-    getServices(): Service[] {
+    getServices(): (Service.Text | Service.TextChem)[] {
         return [Service.Text];
     }