+++ src/RcsbSeacrh/FacetStore/FacetStoreInterface.ts
@@ -2,7 +2,7 @@ import {FacetMemberInterface} from "./FacetMemberInterface";
 import {ReturnType, Service} from "@rcsb/rcsb-api-tools/build/RcsbSearch/Types/SearchEnums";
 
 export interface FacetStoreInterface {
-    getServices(): Service[];
+    getServices(): (Service.Text | Service.TextChem)[];
     getFacetService(service: Service|"all"): FacetMemberInterface[];
     readonly facetLayoutGrid: string[];
     readonly returnType:ReturnType;