+++ src/RcsbSeacrh/FacetStore/UniprotGroupFacetStore.ts
@@ -51,7 +51,7 @@ class UniprotGroupFacetStore implements FacetStoreInterface{
         cloneDeep<FacetMemberInterface>(CHEM_COMP_FACET)
     ];
 
-    getServices(): Service[] {
+    getServices(): (Service.Text | Service.TextChem)[] {
         return [Service.Text];
     }