+++ src/RcsbSeacrh/SearchQueryTools.ts
@@ -1,15 +1,17 @@
-import {SearchQueryType} from "./SearchRequestProperty";
 import {
     LogicalOperator,
-    Operator, RelevanceScoreRankingOption, ReturnType,
+    RelevanceScoreRankingOption,
+    ReturnType,
     ScoringStrategy,
-    Service, SortDirection,
+    Service,
+    SortDirection,
     Type
 } from "@rcsb/rcsb-api-tools/build/RcsbSearch/Types/SearchEnums";
-import {RcsbSearchMetadata} from "@rcsb/rcsb-api-tools/build/RcsbSearch/Types/SearchMetadata";
+import {RcsbSearchAttributeType, RcsbSearchMetadata} from "@rcsb/rcsb-api-tools/build/RcsbSearch/Types/SearchMetadata";
 import {
-    SearchQuery,
-    AttributeTextQueryParameters, ResultsContentType
+    AttributeTextQueryParameters,
+    ResultsContentType,
+    SearchQuery
 } from "@rcsb/rcsb-api-tools/build/RcsbSearch/Types/SearchQueryInterface";
 import {FacetStoreInterface} from "./FacetStore/FacetStoreInterface";
 import {cloneDeep} from 'lodash';
@@ -17,59 +19,35 @@ import {GroupProvenanceId} from "@rcsb/rcsb-api-tools/build/RcsbDw/Types/DwEnums
 import {depositionGroupFacetStore} from "./FacetStore/DepositionGroupFacetStore";
 import {sequenceGroupFacetStore} from "./FacetStore/SequenceGroupFacetStore";
 import {uniprotGroupFacetStore} from "./FacetStore/UniprotGroupFacetStore";
+import {SearchQueryType} from "@rcsb/rcsb-search-tools/lib/SearchQueryTools/SearchQueryInterfaces";
+import {
+    buildAttributeQuery,
+    buildRequestFromCombinedSearchQuery, buildRequestFromSearchQuery
+} from "@rcsb/rcsb-search-tools/lib/SearchQueryTools/SearchQueryTools";
 
 export namespace SearchQueryTools {
 
-    export function searchGroupQuery(groupProvenance: GroupProvenanceId, groupId: string, service?: Service): SearchQueryType {
+    export function searchGroupQuery(groupProvenance: GroupProvenanceId, groupId: string, service?: Service.Text | Service.TextChem): SearchQueryType {
         const groupSearchAttr: GroupSearchAttribute = getSearchAttribute(groupProvenance);
-        return {
-            type: Type.Terminal,
-            service: service ?? Service.Text,
-            parameters: {
-                attribute: groupSearchAttr.path,
-                negation: false,
-                operator: groupSearchAttr.operator.ExactMatch,
-                value: groupId
-            }
-        }
+        return buildAttributeQuery({
+            attribute: groupSearchAttr.path,
+            value: groupId,
+            operator: groupSearchAttr.operator.ExactMatch,
+            service:  service ?? Service.Text
+        });
     }
 
-    export function addGroupNodeToSearchQuery(groupProvenanceId: GroupProvenanceId, groupId: string, searchQuery: SearchQueryType, service?: Service): SearchQueryType {
+    export function addGroupNodeToSearchQuery(groupProvenanceId: GroupProvenanceId, groupId: string, searchQuery: SearchQueryType, service?: Service.Text | Service.TextChem): SearchQueryType {
         return addNodeToSearchQuery(searchGroupQuery(groupProvenanceId, groupId, service), searchQuery);
     }
 
-    export function buildAttributeSearchQuery(attribute: string, value: AttributeTextQueryParameters['value'], operator: Operator, searchQuery: SearchQueryType, returnType: ReturnType, service: Service.Text | Service.TextChem, negation: boolean = false): SearchQuery {
-        return {
-            query: addNewNodeToAttributeSearchQuery(attribute, value, operator, searchQuery, service, negation),
-            return_type: returnType,
-            request_options: {
-                paginate: {
-                    start: 0,
-                    rows: 25
-                },
-                scoring_strategy: ScoringStrategy.Combined,
-                sort: [
-                    {
-                        sort_by: RelevanceScoreRankingOption.Score,
-                        direction: SortDirection.Desc
-                    }
-                ]
-            }
-        }
-    }
-
     export function buildNodeSearchQuery(node: SearchQueryType, searchQuery: SearchQueryType, returnType: ReturnType, resultsContentType:ResultsContentType, logicalOperator = LogicalOperator.And): SearchQuery {
-        return {
-            query: {
-                type: Type.Group,
-                logical_operator: logicalOperator,
-                nodes: [
-                    searchQuery,
-                    node
-                ]
-            },
-            return_type: returnType,
-            request_options: {
+        return buildRequestFromCombinedSearchQuery(
+            searchQuery,
+            node,
+            returnType,
+            logicalOperator,
+            {
                 paginate: {
                     start: 0,
                     rows: 25
@@ -83,14 +61,14 @@ export namespace SearchQueryTools {
                 ],
                 results_content_type: resultsContentType
             }
-        }
+        );
     }
 
     export function buildSearchQuery(searchQuery: SearchQueryType, returnType: ReturnType): SearchQuery {
-        return {
-            query: searchQuery,
-            return_type: returnType,
-            request_options: {
+        return buildRequestFromSearchQuery(
+            searchQuery,
+            returnType,
+            {
                 paginate: {
                     start: 0,
                     rows: 25
@@ -102,25 +80,21 @@ export namespace SearchQueryTools {
                         direction: SortDirection.Desc
                     }
                 ]
-            }
-        }
+            });
     }
 
-    export function addNewNodeToAttributeSearchQuery(attribute: string, value: AttributeTextQueryParameters['value'], operator: Operator, searchQuery: SearchQueryType, service: Service.Text | Service.TextChem, negation: boolean = false): SearchQueryType {
+    export function addNewNodeToAttributeSearchQuery(attribute: RcsbSearchAttributeType, value: AttributeTextQueryParameters['value'], operator: AttributeTextQueryParameters["operator"], searchQuery: SearchQueryType, service: Service.Text | Service.TextChem, negation: boolean = false): SearchQueryType {
         return addNodeToSearchQuery(searchAttributeQuery(attribute, value, operator, service, negation), searchQuery);
     }
 
-    export function searchAttributeQuery(attribute: string, value: AttributeTextQueryParameters['value'], operator: Operator, service: Service.Text | Service.TextChem, negation: boolean = false): SearchQueryType {
-        return {
-            type: Type.Terminal,
-            service: service,
-            parameters: {
-                attribute: attribute,
-                negation: negation,
-                operator: operator as any,
-                value: value
-            }
-        }
+    export function searchAttributeQuery(attribute: RcsbSearchAttributeType, value: AttributeTextQueryParameters['value'], operator: AttributeTextQueryParameters["operator"], service: Service.Text | Service.TextChem, negation: boolean = false): SearchQueryType {
+        return buildAttributeQuery({
+            attribute,
+            value,
+            operator,
+            service,
+            negation
+        });
     }
 
     export function addNodeToSearchQuery(node: SearchQueryType, searchQuery: SearchQueryType, logicalOperator = LogicalOperator.And): SearchQueryType {
@@ -155,8 +129,8 @@ export namespace SearchQueryTools {
     }
 
     type GroupSearchAttribute =
-        (typeof RcsbSearchMetadata.RcsbEntryGroupMembership.GroupId)
-        | (typeof RcsbSearchMetadata.RcsbPolymerEntityGroupMembership.GroupId);
+        (typeof RcsbSearchMetadata.RcsbEntryGroupMembership.GroupId) |
+        (typeof RcsbSearchMetadata.RcsbPolymerEntityGroupMembership.GroupId);
 
     function getSearchAttribute(groupProvenanceId: GroupProvenanceId): GroupSearchAttribute {
         switch (groupProvenanceId) {