+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 [Semantic Versioning](https://semver.org/)
 
+## [5.0.8] - 2023-05-30
+### Improvement
+- New type `ChartObjectIdType` to identify included and excluded distribution areas in group summary pages
+- Strict typed `FacetMemberInterface.attribute` and `RcsbChartInterface.attribute` to `RcsbSearchAttributeType`
+### Refactoring
+- Using rcsb-search-tools
+- Search requests use rcsb-search-tools methods
+  - `FacetTools.ts`
+  - `SearchQueryTools.ts`
+### Dependency update
+- rcsb-charts v0.1.4
+
 ## [5.0.7] - 2023-05-16
 ### Bug fix
 - Uniprot unknown annotation type bug fixed